Added directional validity checks
I've added the ability to check whether a given direction is valid in the Player and Room classes. I did this by adding a temporary direction class with a valid_directions list that can be check against.

In addition, I've created the directions as dicts containing the direction name and a vector describing which way the direction should move us.

This class will be replaced when a list of valid directions can be pushed into the world database on a per-game basis. Some version of this may remain here to both prevent changes to the existing classes and to keep some hard-wired values from which to build a DB or to provide a fallthrough for an incomplete DB.