World class partially works but will be rewritten
I've been convinced that as it is, I can't do arbitrarily large levels with this design.

I'm going to rewrite it for a 500^3 limit. Hopefully this will get rid of the try: soup in ProcessWorld. I want to commit it in case I need look back at it at some point.